requirements.txt
flask==2.3.2
tensorflow==2.12.0
numpy==1.24.3
pillow==9.5.0
werkzeug==2.3.7